int to long

+++ src/main/java/com/websystique/springmvc/dao/UserDao.java
@@ -7,7 +7,7 @@ import com.websystique.springmvc.model.User;
 
 public interface UserDao {
 
-	User findById(int id);
+	User findById(long id);
 	
 	User findBySSO(String sso);
