+++ src/main/java/com/websystique/springmvc/dao/UserProfileDao.java
@@ -11,5 +11,5 @@ public interface UserProfileDao {
 	
 	UserProfile findByType(String type);
 	
-	UserProfile findById(int id);
+	UserProfile findById(long id);
 }
