+++ src/main/java/com/websystique/springmvc/service/UserService.java
@@ -7,7 +7,7 @@ import com.websystique.springmvc.model.User;
 
 public interface UserService {
 	
-	User findById(int id);
+	User findById(long id);
 	
 	User findBySSO(String sso);
 	
@@ -19,6 +19,6 @@ public interface UserService {
 
 	List<User> findAllUsers(); 
 	
-	boolean isUserSSOUnique(Integer id, String sso);
+	boolean isUserSSOUnique(Long id, String sso);
 
 }
